backend/requirements.txt
# backend\requirements.txt

fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
pandas>=2.1
numpy>=1.25
prophet>=1.1.5
xgboost>=2.0
scikit-learn>=1.3
sqlmodel>=0.0.8
pyyaml>=6.0
requests>=2.31
google-generativeai>=0.3